short -> int conversion
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1049 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ dig.java
@@ -177,7 +177,7 @@ main(String argv[]) throws IOException {
 			usage();
 	}
 
-	rec = Record.newRecord(name, (short) type, (short) dclass);
+	rec = Record.newRecord(name, type, dclass);
 	query = Message.newQuery(rec);
 	if (opt != null)
 		query.addRecord(opt, Section.ADDITIONAL);
